requirements.txt
beautifulsoup4
requests
pandas
streamlit
pinecone-client>=2.2.1
python-dotenv
openai